Move release gate test support into test libraries
The module is only imported by tests.\nMoving it under source/test-libraries keeps the test harness readable and removes it from Stryker's mutate set.
